+++ src/errors/al-sc-exist.error.ts
@@ -1,0 +1,7 @@
+export class ALSCExistError extends Error {
+    constructor(message: string = 'AL folder already exists') {
+      super(message);
+  
+      this.name = 'ALSCExistError';
+    }
+  }