+++ CHANGELOG.md
@@ -1,9 +1,25 @@
 # Change Log
 
-All notable changes to the "alfolderstructure" extension will be documented in this file.
-
-Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
+All notable changes to the **AL Structure Creator** extension will be documented in this file.
 
 ## [Unreleased]
 
-- Initial release
+### [0.1.0]
+
+#### Added
+- New command to move/**reorganize** application objects to the corresponding application folders.
+- New configurations to be used in the new command.
+- Add new folder validations.
+
+#### Removed
+- Configuration `ProjectFolderOption`.
+
+### 0.0.2
+
+#### Added
+
+- New command to create an application object folder individually.
+
+### 0.0.1
+
+Initial release.